Always use public path
